requirements.txt
paramiko==3.5.1
pillow==12.0.0
pydantic-settings==2.5.2
pyinstaller==6.16.0
pyside6==6.10.0
python-dotenv==1.0.1
send2trash==1.8.3
watchdog==5.0.3